split services from components folder
