+++ packages/sitecore-jss-angular/src/services/data-resolver-factory.spec.ts
