+++ packages/sitecore-jss-angular/src/services/jss-can-activate-error.ts
